backend/medical_report_analyzer/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
PyMuPDF
openai
python-multipart
